+++ src/core/client/views/creator/html/components/makeup.js
@@ -50,7 +50,7 @@ const MakeupComponent = Vue.component('makeup', {
                 return;
             }
 
-            alt.emit('creator:PlaySound', 'TOGGLE_ON', 'HUD_FRONTEND_DEFAULT_SOUNDSET');
+            alt.emit('play:Sound', 'TOGGLE_ON', 'HUD_FRONTEND_DEFAULT_SOUNDSET');
         },
         getLocale(id) {
             return this.locales.makeupComponent.ids[id]